pyproject.toml
[tool.poetry]
name = "phonometrics"
version = "0.0.1"
description = ""
authors = ["Alexander Bondarenko <sascha.bondarenko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
torch = "^2.5.1"
torchaudio = "^2.5.1"
transformers = "^4.46.2"
soundfile = "^0.12.1"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
python-multipart = "^0.0.17"
pyyaml = "^6.0.2"
openai = "^1.54.3"
ffmpeg-python = "^0.2.0"
llvmlite = "0.43.0"
numba = "0.60.0"
openai-whisper = "^20240930"
python-dotenv = "^1.0.1"
praat-parselmouth = "^0.4.5"
matplotlib = "^3.9.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
black = "^24.10.0"
flake8 = "^7.1.1"
mypy = "^1.13.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
force_single_line = true
atomic = true
include_trailing_comma = true
lines_after_imports = 2
lines_between_types = 1
use_parentheses = true
skip_glob = ["*/setup.py"]

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | docs
  | tests/.*/setup.py
  | phonometrics/transcription/phonemes/*.*
)/
'''

[tool.mypy]
exclude = ['phonometrics/transcription/phonemes/*.*']


[tool.pytest.ini_options]
markers = [
    "integration: requires running docker container"
]